fetcher: common -> scraper
