+++ kifi-backend/modules/scraper/app/com/keepit/scraper/HttpFetcher.scala
@@ -4,7 +4,6 @@ import org.joda.time.DateTime
 import com.keepit.model.HttpProxy
 import org.apache.http.protocol.{BasicHttpContext, HttpContext}
 import org.apache.http._
-import com.keepit.common.net.URI
 import com.keepit.common.logging.Logging
 import org.apache.http.config.RegistryBuilder
 import org.apache.http.conn.socket.{PlainConnectionSocketFactory, ConnectionSocketFactory}
@@ -23,7 +22,7 @@ import com.keepit.common.time._
 import com.keepit.common.performance.{timing, timingWithResult}
 import java.util.concurrent.{ThreadFactory, TimeUnit, Executors, ConcurrentLinkedQueue}
 import scala.ref.WeakReference
-import play.api.{Logger, Play}
+import play.api.Play
 import Play.current
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import java.util.concurrent.atomic.{AtomicInteger, AtomicReference}
@@ -41,30 +40,6 @@ trait HttpFetcher { // todo(ray): move to scraper
   def fetch(url: String, ifModifiedSince: Option[DateTime] = None, proxy: Option[HttpProxy] = None)(f: HttpInputStream => Unit): HttpFetchStatus
   def close()
 }
-
-case class HttpFetchStatus(statusCode: Int, message: Option[String], context: HttpContext) {
-  def destinationUrl = Option(context.getAttribute("scraper_destination_url").asInstanceOf[String])
-  def redirects = Option(context.getAttribute("redirects").asInstanceOf[Seq[HttpRedirect]]).getOrElse(Seq.empty[HttpRedirect])
-}
-
-case class HttpRedirect(statusCode: Int, currentLocation: String, newDestination: String) {
-  def isPermanent = (statusCode == HttpStatus.SC_MOVED_PERMANENTLY)
-  def isAbsolute = URI.isAbsolute(currentLocation) && URI.isAbsolute(newDestination)
-  def isLocatedAt(url: String) = (currentLocation == url)
-}
-
-object HttpRedirect {
-  def withStandardizationEffort(statusCode: Int, currentLocation: String, newDestination: String): HttpRedirect = HttpRedirect(statusCode, currentLocation, URI.absoluteUrl(currentLocation, newDestination).getOrElse(newDestination))
-
-  import play.api.libs.functional.syntax._
-  import play.api.libs.json._
-  implicit val format = (
-    (__ \ 'statusCode).format[Int] and
-    (__ \ 'currentLocation).format[String] and
-    (__ \ 'newDestination).format[String]
-  )(HttpRedirect.apply _, unlift(HttpRedirect.unapply))
-}
-
 
 class HttpFetcherImpl(val airbrake:AirbrakeNotifier, userAgent: String, connectionTimeout: Int, soTimeOut: Int, trustBlindly: Boolean, schedulingProperties:SchedulingProperties) extends HttpFetcher with Logging with ScraperUtils {
   val cm = if (trustBlindly) {
